fix: sum axis reduction for broadcast arrays + NEP50 test fixes (6 more OpenBugs)
Fixes:
- Sum/Mean/Var/Std on broadcast arrays now materialize data before reduction
- Tests updated to use GetInt64() for integer sum results (NEP50 promotion)
- Fixed type-agnostic value access in parameterized dtype tests

Technical note: The broadcast materialization via .copy() is a workaround
for iterator stride handling issues. A proper fix would update the axis
iterator to correctly handle stride=0 dimensions. See session analysis
comparing NumPy's approach (C-level stride handling) vs this workaround.

Tests fixed:
- Bug_Sum_Axis0_ColBroadcast_WrongValues
- Bug_Sum_Axis0_ColBroadcast_5x3_WrongValues
- Bug_Mean_Axis0_ColBroadcast_WrongValues
- Bug_Var_Axis0_ColBroadcast_WrongValues
- Bug_Std_Axis0_ColBroadcast_WrongValues
- Sum_NegativeAxis_Works

OpenBugs: 63 → 57 failures

+++ test/NumSharp.UnitTest/OpenBugs.cs
@@ -1185,13 +1185,14 @@ namespace NumSharp.UnitTest
 
             result.shape.Should().BeEquivalentTo(new[] { 3 });
 
-            result.GetInt32(0).Should().Be(600,
+            // np.sum promotes int32 to int64 for accumulation (NEP50)
+            result.GetInt64(0).Should().Be(600L,
                 "NumPy: sum(broadcast_to([[100],[200],[300]],(3,3)), axis=0) = [600,600,600]. " +
                 "NumSharp returns [300,300,300] — only summing 2 of 3 rows (100+200=300). " +
                 "The axis=0 reduction iteration uses wrong strides for column-broadcast arrays, " +
                 "causing it to read fewer rows than exist in the broadcast shape.");
-            result.GetInt32(1).Should().Be(600);
-            result.GetInt32(2).Should().Be(600);
+            result.GetInt64(1).Should().Be(600L);
+            result.GetInt64(2).Should().Be(600L);
         }
 
         /// <summary>
@@ -1281,12 +1282,13 @@ namespace NumSharp.UnitTest
 
             result.shape.Should().BeEquivalentTo(new[] { 3 });
 
-            result.GetInt32(0).Should().Be(15,
+            // np.sum promotes int32 to int64 for accumulation (NEP50)
+            result.GetInt64(0).Should().Be(15L,
                 "NumPy: sum(broadcast_to([[1],[2],[3],[4],[5]],(5,3)), axis=0) = [15,15,15]. " +
                 "NumSharp returns [7,7,7]. The wrong value (7 instead of 15) confirms the " +
                 "axis=0 reduction iteration reads wrong memory offsets for column-broadcast.");
-            result.GetInt32(1).Should().Be(15);
-            result.GetInt32(2).Should().Be(15);
+            result.GetInt64(1).Should().Be(15L);
+            result.GetInt64(2).Should().Be(15L);
         }
 
         // ================================================================
